+++ Thaw/MenuBar/Search/MenuBarSearchPanel.swift
@@ -866,7 +866,7 @@ private struct SettingsButton: View {
 
     var body: some View {
         Button(action: action) {
-            Image(.iceCubeStroke)
+            Image(.tidybarStroke)
                 .resizable()
                 .aspectRatio(contentMode: .fit)
                 .foregroundStyle(.secondary)